Update to Clash 1.6+
We now include the testsuite source files in clash-compiler directly as
a cabal hs-source-dir, and that code has switched to using MemBlob
instead of the `asyncRomFile` hackery. This should be a much better
structure, and so we can stop throwing out old compilation results in
`run-clash.sh` as well.

`clashFunctions.tcl` now reports total wall time spent running test
benches.