Make a physical test bench
A new Clash design, `PhysTB`, and a new Vivado project `synth.tcl`
together form a design that can be loaded on a Digilent Arty-A7-35T to
verify that it is indeed possible to synthesize Xilinx floating point
IP, not just run them in simulation.

`run-clash.sh` and `run-vivado.sh` are consolidated into `run.sh` with
extended features to make use quick and painless.